fix: re-verify sha256 on pfilter cache hits, restrict FFI to 64-bit darwin/linux, pin CI to Go 1.26
