+++ pfilter/pfilter_test.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"os"
 	"testing"
-	"unsafe"
 
 	"github.com/hoophq/alcatraz/analyzer"
 	"github.com/hoophq/alcatraz/entities"
@@ -31,27 +30,6 @@ func newFakeEngine(c classifier, cfg Config) *Engine {
 		cfg.LabelMapping = DefaultConfig("x.gguf").LabelMapping
 	}
 	return &Engine{classifier: c, cfg: cfg}
-}
-
-func TestPfEntityLayout(t *testing.T) {
-	// The FFI layer reads pf_entity arrays by pointer arithmetic; this pins
-	// the assumed C struct layout (see pf.h).
-	var e pfEntity
-	if size := unsafe.Sizeof(e); size != 24 {
-		t.Errorf("sizeof(pfEntity) = %d, want 24", size)
-	}
-	if off := unsafe.Offsetof(e.start); off != 0 {
-		t.Errorf("offsetof(start) = %d, want 0", off)
-	}
-	if off := unsafe.Offsetof(e.end); off != 4 {
-		t.Errorf("offsetof(end) = %d, want 4", off)
-	}
-	if off := unsafe.Offsetof(e.score); off != 8 {
-		t.Errorf("offsetof(score) = %d, want 8", off)
-	}
-	if off := unsafe.Offsetof(e.label); off != 16 {
-		t.Errorf("offsetof(label) = %d, want 16", off)
-	}
 }
 
 func TestDefaultConfigSupportedEntities(t *testing.T) {
