+++ pfilter/pool_test.go
@@ -1,5 +1,3 @@
-//go:build darwin || linux
-
 package pfilter
 
 import (
